orchestrator/requirements.txt
solana>=0.36.0
solders>=0.26.0
google-cloud-compute>=1.18.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=8.0
pytest-asyncio>=0.23
